refactor: split HexMapDisplay.tsx into sub-components
- Created 7 focused files (89-341 lines each)
- Extracted rendering helpers (renderHelpers.ts, 189 lines)
- Extracted interaction hook (useMapInteraction.ts, 184 lines)
- Created sub-components (HexCell, UnitToken, Overlays)
- Main component orchestrates sub-components (341 lines)
- All files under FILE_MODULARITY_SPEC thresholds
- Zero behavior changes (structural split only)
