+++ src/components/gameplay/HexMapDisplay/index.tsx
@@ -1,0 +1,2 @@
+export { HexMapDisplay, default } from './HexMapDisplay';
+export type { HexMapDisplayProps } from './HexMapDisplay';
